Fix folder creation and modification

+++ app/routes/controllers/folders/delete.py
@@ -1,14 +1,21 @@
-from fastapi import HTTPException, Depends
+from fastapi import HTTPException
 from app.services.connection import mail_db
 
-async def delete_folder(id: int):
-    
+async def delete_folder(id: int, user_id: str):
+    user_db = mail_db.fetch_one(
+        sql='SELECT * FROM users WHERE id = UUID_TO_BIN(%s)',
+        params=(user_id,)
+    )
+    if not user_db:
+        raise HTTPException(status_code=404, detail='User not found')
+
     folder_db = mail_db.fetch_one(
         sql='SELECT * FROM folders WHERE id = %s',
         params=(id,)
     )
     if not folder_db:
         raise HTTPException(status_code=404, detail='Folder not found')
+
     try:
         mail_db.execute(
             sql='DELETE FROM folders WHERE id = %s',
@@ -17,8 +24,8 @@ async def delete_folder(id: int):
     except Exception as e:
         print(e)
         raise HTTPException(status_code=500, detail='Ocurrio un error al eliminar la carpeta')
-    
+
     return {
         'message': 'Folder deleted successfully',
-        'folder_id': folder_db
+        'folder_id': id
         }
